refactor: remove form tour settings
