+++ frappe/desk/doctype/form_tour/form_tour.py
@@ -1,6 +1,8 @@
 # Copyright (c) 2021, Frappe Technologies and contributors
 # License: MIT. See LICENSE
 
+import json
+
 import frappe
 from frappe.model.document import Document
 from frappe.modules.export_file import export_to_files
@@ -32,40 +34,12 @@ class FormTour(Document):
 
 	def on_update(self):
 		frappe.cache().delete_key("bootinfo")
-		if self.ui_tour:
-			form_tour_settings = frappe.get_doc("Form Tour Settings", "Form Tour Settings")
-			in_settings = False
-			child_index = 0
-			for tour in form_tour_settings.form_tours:
-				if tour.form_tour == self.name:
-					in_settings = True
-					child_index = tour.idx
-					form_tour_settings.remove(tour)
-			if not in_settings:
-				child_index = len(form_tour_settings.form_tours) + 1
-			child = frappe.new_doc("Form Tour Settings Item")
-			child.update(
-				{
-					"idx": child_index,
-					"form_tour": self.name,
-					"parent": "Form Tour Settings",
-					"parentfield": "form_tours",
-					"parenttype": "Form Tour Settings",
-				}
-			)
-			child.save()
-			form_tour_settings.form_tours.insert(child_index, child)
-			form_tour_settings.save()
+
 		if frappe.conf.developer_mode and self.is_standard:
 			export_to_files([["Form Tour", self.name]], self.module)
 
 	def on_trash(self):
-		if self.ui_tour:
-			form_tour_settings = frappe.get_doc("Form Tour Settings", "Form Tour Settings")
-			for tour in form_tour_settings.form_tours:
-				if tour.form_tour == self.name:
-					form_tour_settings.remove(tour)
-			form_tour_settings.save()
+		frappe.cache().delete_key("bootinfo")
 
 
 @frappe.whitelist()
@@ -76,3 +50,31 @@ def reset_tour(tour_name):
 		onboarding_status.pop(tour_name, None)
 		user_doc.onboarding_status = frappe.as_json(onboarding_status)
 		user_doc.save()
+
+
+@frappe.whitelist()
+def update_user_status(value, step):
+	from frappe.utils.telemetry import capture
+
+	step = frappe.parse_json(step)
+	tour = frappe.parse_json(value)
+
+	capture(
+		frappe.scrub(f"{step.parent}_{step.title}"),
+		app="frappe_ui_tours",
+		properties={"is_completed": tour.is_completed},
+	)
+	frappe.db.set_value(
+		"User", frappe.session.user, "onboarding_status", value, update_modified=False
+	)
+
+	frappe.cache().hdel("bootinfo", frappe.session.user)
+
+
+def get_onboarding_ui_tours():
+	if not frappe.get_system_settings("enable_onboarding"):
+		return []
+
+	ui_tours = frappe.get_all("Form Tour", filters={"ui_tour": 1}, fields=["page_route", "name"])
+
+	return [[tour.name, json.loads(tour.page_route)] for tour in ui_tours]
