--- frappe/desk/doctype/form_tour_settings/form_tour_settings.json
@@ -1,51 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "creation": "2023-05-11 18:07:26.879273",
- "default_view": "List",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "onboarding_tours",
-  "form_tours"
- ],
- "fields": [
-  {
-   "fieldname": "form_tours",
-   "fieldtype": "Table",
-   "label": "Form Tours",
-   "options": "Form Tour Settings Item"
-  },
-  {
-   "default": "\"[]\"",
-   "fieldname": "onboarding_tours",
-   "fieldtype": "Small Text",
-   "hidden": 1,
-   "label": "Onboarding Tours"
-  }
- ],
- "index_web_pages_for_search": 1,
- "issingle": 1,
- "links": [],
- "modified": "2023-05-24 12:05:52.676242",
- "modified_by": "Administrator",
- "module": "Desk",
- "name": "Form Tour Settings",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "print": 1,
-   "read": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": []
-}
