--- frappe/desk/doctype/form_tour_settings/test_form_tour_settings.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2023, Frappe Technologies and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestFormTourSettings(FrappeTestCase):
-	pass
