--- frappe/desk/doctype/form_tour_settings_item/form_tour_settings_item.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2023, Frappe Technologies and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-
-class FormTourSettingsItem(Document):
-	pass
