+++ frappe/utils/install.py
@@ -14,8 +14,6 @@ def before_install():
 	frappe.reload_doc("core", "doctype", "doctype_action")
 	frappe.reload_doc("core", "doctype", "doctype_link")
 	frappe.reload_doc("desk", "doctype", "form_tour_step")
-	frappe.reload_doc("desk", "doctype", "form_tour_settings")
-	frappe.reload_doc("desk", "doctype", "form_tour_settings_item")
 	frappe.reload_doc("desk", "doctype", "form_tour")
 	frappe.reload_doc("core", "doctype", "doctype")
